saving before album controller code
