+++ mobile_mentor_ios_code_challenge.xcworkspace/xcuserdata/elidangerfield.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,7 +10,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "mobile_mentor_ios_code_challenge/ViewControllers/SearchViewController.swift"
-            timestampString = "580327342.425711"
+            timestampString = "580328016.837955"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "30"
